--- java_oop_reference/.gitignore
@@ -1,3 +1,0 @@
-.idea
-out/
-*.iml
